:tada: feat: update arrest reports tab
