+++ apps/client/src/pages/officer/supervisor/citizen-logs/index.tsx
@@ -9,16 +9,17 @@ import { Title } from "components/shared/Title";
 import { Permissions } from "@snailycad/permissions";
 import { TabList } from "components/shared/TabList";
 import { CitizenLogsTab } from "components/leo/citizen-logs/citizen-logs-tab";
-import { ArrestReportsTab } from "components/leo/citizen-logs/ArrestReportsTab";
+import { ArrestReportsTab } from "components/leo/citizen-logs/arrest-reports-tab";
 import { useFeatureEnabled } from "hooks/useFeatureEnabled";
-import type { GetManageRecordLogsData } from "@snailycad/types/api";
+import type { GetManagePendingArrestReports, GetManageRecordLogsData } from "@snailycad/types/api";
 
 export type CitizenLog = RecordLog & { citizen: Citizen };
 interface Props {
   citizens: GetManageRecordLogsData;
+  arrestReports: GetManagePendingArrestReports;
 }
 
-export default function CitizenLogs({ citizens }: Props) {
+export default function CitizenLogs(props: Props) {
   const { CITIZEN_RECORD_APPROVAL } = useFeatureEnabled();
 
   const t = useTranslations("Leo");
@@ -40,8 +41,8 @@ export default function CitizenLogs({ citizens }: Props) {
       <Title>{t("citizenLogs")}</Title>
 
       <TabList tabs={TABS}>
-        <CitizenLogsTab citizens={citizens} />
-        {CITIZEN_RECORD_APPROVAL ? <ArrestReportsTab logs={citizens} /> : null}
+        <CitizenLogsTab citizens={props.citizens} />
+        {CITIZEN_RECORD_APPROVAL ? <ArrestReportsTab arrestReports={props.arrestReports} /> : null}
       </TabList>
     </Layout>
   );
@@ -49,13 +50,15 @@ export default function CitizenLogs({ citizens }: Props) {
 
 export const getServerSideProps: GetServerSideProps = async ({ req, locale }) => {
   const user = await getSessionUser(req);
-  const [citizens] = await requestAll(req, [
-    ["/admin/manage/citizens/records-logs", { citizens: [], totalCount: 0 }],
+  const [citizens, arrestReports] = await requestAll(req, [
+    ["/admin/manage/records-logs", { citizens: [], totalCount: 0 }],
+    ["/admin/manage/pending-arrest-reports", { arrestReports: [], totalCount: 0 }],
   ]);
 
   return {
     props: {
       session: user,
+      arrestReports,
       citizens,
       messages: {
         ...(await getTranslations(["leo", "common"], user?.locale ?? locale)),
